+++ src/main/java/net/silentchaos512/gear/api/parts/IGearPart.java
@@ -67,6 +67,10 @@ public interface IGearPart {
         return normalLossRate;
     }
 
+    default boolean isCraftingAllowed(@Nullable GearType gearType) {
+        return true;
+    }
+
     IPartDisplay getDisplayProperties(PartData part, ItemStack gear, int animationFrame);
 
     @Nullable